% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDependencies.R
\name{checkDependencies}
\alias{checkDependencies}
\title{Checks whether the method is implemented and whether all dependencies are
installed}
\usage{
checkDependencies(method)
}
\arguments{
\item{method}{Method to use.}
}
\description{
Checks whether the method is implemented and whether all dependencies are
installed
}
