\name{CompareCausalNetworks-package}
\alias{CompareCausalNetworks-package}
\alias{CompareCausalNetworks}
\docType{package}
\title{
Compare estimates of causal graphs using a unified interface to various methods
}
\description{
Provides a unified interface to various causal graph estimation methods.
}
\details{
\tabular{ll}{
Package: \tab CompareCausalNetworks\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2015-06-17\cr
License: \tab GPL\cr
}
The causal graphs can be estimated with function \code{\link{getParents}} 
and a stability-selection version is available at \code{\link{getParentsStable}}. %The resulting graphs can be plotted for example with \code{\link{plotGraph}}.
}
\author{
Christina Heinze <heinze@stat.math.ethz.ch>, Nicolai Meinshausen <meinshausen@stat.math.ethz.ch>
Maintainer: Christina Heinze <heinze@stat.math.ethz.ch>, Nicolai Meinshausen <meinshausen@stat.math.ethz.ch>
}