% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData_unknownShiftInterventions.R
\docType{data}
\name{simData_unknownShiftInterventions}
\alias{simData_unknownShiftInterventions}
\title{Data from a causal cyclic model with shift interventions}
\format{A list created by \code{\link[backShift]{simulateInterventions}}. 
All inputs are contained in the list element \code{configs}. For details
see the help page of \code{\link[backShift]{simulateInterventions}}.}
\usage{
simData_unknownShiftInterventions
}
\description{
A dataset to run the tests.
}
\keyword{datasets}
