% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData_wInterventions.R
\docType{data}
\name{simDataInv}
\alias{simDataInv}
\title{Data from a causal model with interventions}
\format{A list created by \code{\link{simulateInterventions}}. 
All inputs are contained in the list element \code{configs}. For details
see the help page of \code{\link{simulateInterventions}}.}
\usage{
simDataInv
}
\description{
A dataset to run the tests.
}
\keyword{datasets}
