% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilitySelectionChecks.R
\name{stabilitySelectionChecks}
\alias{stabilitySelectionChecks}
\title{Check the validity of the stability selection inputs.}
\usage{
stabilitySelectionChecks(p, q, nodewise, threshold, sampleSettings,
  sampleObservations)
}
\arguments{
\item{p}{Number of variables}

\item{q}{Stability selection parameter}

\item{nodewise}{Do stability selection nodewise?}

\item{threshold}{Threshold for stability selection.}

\item{sampleSettings}{The proportion of unique settings to resample for each 
resample; has to be in [0,1].}

\item{sampleObservations}{The fraction of all samples to retain when 
subsampling (no replacement); has to be in [0,1].}
}
\description{
Check the validity of the stability selection inputs.
}
