% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adjustOptions.R
\name{adjustOptions}
\alias{adjustOptions}
\title{Helper function to change default options}
\usage{
adjustOptions(availableOptions, optionsToSet)
}
\arguments{
\item{availableOptions}{Options provided by function.}

\item{optionsToSet}{Options to be changed to user-provided values.}
}
\description{
Helper function to change default options
}

