% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateInterventions}
\alias{simulateInterventions}
\title{Simulate data of a causal (possibly cyclic model) under interventions.}
\usage{
simulateInterventions(n, p, df, rhoNoise, snrPar, sparse, doInterv, numberInt,
  strengthInt, cyclic, strengthCycle, modelMis = FALSE, modelMisPar = 1,
  seed = 1)
}
\arguments{
\item{n}{Number of observations.}

\item{p}{Number of variables.}

\item{df}{Degrees of freedom in t-distribution of noise.}

\item{rhoNoise}{Correlation between noise terms. Set to 0 for independent noise.}

\item{snrPar}{SNR parameter: steers what proportion of the variance stems from 
the signal resp.\ from the noise ($SNR = (1-\code{snrPar})/\code{snrPar}$)}

\item{sparse}{Probability that an entry i,j in adjacency matrix is 1.}

\item{doInterv}{Set to TRUE if interventions should be do-interventions; otherwise
noise interventions are generated.}

\item{numberInt}{Number of interventions.}

\item{strengthInt}{Regulates the strength of the interventions.}

\item{cyclic}{Set to TRUE is resulting graph should contain a cycle.}

\item{strengthCycle}{Steers strength of feedback.}

\item{modelMis}{Add a model misspecification that applies \code{tanh(modelMisPar*x)/modelMisPar)}
morginally to each variable after having generated X from the causal DAG.}

\item{modelMisPar}{Parameter steering the strength of the model misspecification.}

\item{seed}{Random seed.}
}
\value{
A list with the following elements: 
\itemize{
  \item \code{X} (nxp)-dimensional data matrix
  \item \code{environment} Indicator of the experiment or the intervention type an 
  observation belongs to. A numeric vector of length n. 
  \item \code{interventions} Location of interventions
  \item \code{whereInt}
  \item \code{noise}
  \item \code{configs} A list with the following elements: 
  \itemize{
  \item ...
  }
}
}
\description{
Simulate data of a causal (possibly cyclic model) under interventions.
}
